# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smit CSC Info API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /credits/me:
    get:
      operationId: getMyCredits
      tags: [tools]
      summary: Get current user's FHD credit balance and Prime status
      responses:
        "200":
          description: Credit balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditBalance"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tools/remove-bg-fhd:
    post:
      operationId: removeBackgroundFhd
      tags: [tools]
      summary: Premium FHD background removal via Remove.bg (Prime only, 1 credit)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
                bgColor:
                  type: ["string", "null"]
                  description: Optional hex color (e.g. ffffff). Omit for transparent.
      responses:
        "200":
          description: PNG image with background removed
          content:
            image/png:
              schema:
                type: string
                format: binary
          headers:
            X-Credits-Remaining:
              schema:
                type: integer
              description: Credits left after this call
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "402":
          description: No credits or not Prime
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Premium engine unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /content:
    get:
      operationId: getContent
      tags: [content]
      summary: Get content items (videos, forms, schemes)
      parameters:
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: type
          in: query
          schema:
            type: ["string", "null"]
        - name: isPrime
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: Content list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentItem"
  /content/{id}:
    get:
      operationId: getContentItem
      tags: [content]
      summary: Get a single content item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Content item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /documents:
    get:
      operationId: getDocuments
      tags: [documents]
      summary: List all documents
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: isPrime
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetDocumentsResponse"
  /admin/documents:
    post:
      operationId: adminCreateDocument
      tags: [admin, documents]
      summary: Admin - Create a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentItem"
  /admin/documents/{id}:
    delete:
      operationId: adminDeleteDocument
      tags: [admin, documents]
      summary: Admin - Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    CreditBalance:
      type: object
      properties:
        isPrime:
          type: boolean
        credits:
          type: integer
        monthlyAllowance:
          type: integer
        cycleStart:
          type: ["string", "null"]
        nextRenewal:
          type: ["string", "null"]
      required:
        - isPrime
        - credits
        - monthlyAllowance
...
    ContentItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        titleGu:
          type: ["string", "null"]
        category:
          type: string
        type:
          type: string
        link:
          type: string
        description:
          type: ["string", "null"]
        isPrime:
          type: boolean
        thumbnailUrl:
          type: ["string", "null"]
        youtubeVideoId:
          type: ["string", "null"]
        playlistId:
          type: ["string", "null"]
        playlistTitle:
          type: ["string", "null"]
        publishedAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - title
        - category
        - type
        - link
        - isPrime
        - createdAt
...
    DocumentItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        fileUrl:
          type: string
        fileName:
          type: string
        fileType:
          type: string
        category:
          type: string
        isPrime:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - title
        - fileUrl
        - fileName
        - fileType
        - category
        - isPrime
        - createdAt
    GetDocumentsResponse:
      type: array
      items:
        $ref: "#/components/schemas/DocumentItem"
    CreateDocumentBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        fileUrl:
          type: string
        fileName:
          type: string
        fileType:
          type: string
        category:
          type: string
        isPrime:
          type: boolean
      required:
        - title
        - fileUrl
        - fileName
        - fileType
...